+++ src/gpu/sequenceconversionkernels.cu
@@ -543,8 +543,8 @@ void encodeSequencesTo2BitKernel(
     unsigned int* __restrict__ encodedSequences,
     const char* __restrict__ decodedSequences,
     const int* __restrict__ sequenceLengths,
-    int decodedSequencePitchInBytes,
-    int encodedSequencePitchInInts,
+    size_t decodedSequencePitchInBytes,
+    size_t encodedSequencePitchInInts,
     int numSequences
 ){
     auto group = cg::tiled_partition<groupsize>(cg::this_thread_block());
@@ -614,8 +614,8 @@ void callEncodeSequencesTo2BitKernel(
     unsigned int* d_encodedSequences,
     const char* d_decodedSequences,
     const int* d_sequenceLengths,
-    int decodedSequencePitchInBytes,
-    int encodedSequencePitchInInts,
+    size_t decodedSequencePitchInBytes,
+    size_t encodedSequencePitchInInts,
     int numSequences,
     int groupsize,
     cudaStream_t stream
@@ -653,8 +653,8 @@ void decodeSequencesFrom2BitKernel(
     char* __restrict__ decodedSequences,
     const unsigned int* __restrict__ encodedSequences,
     const int* __restrict__ sequenceLengths,
-    int decodedSequencePitchInBytes,
-    int encodedSequencePitchInInts,
+    size_t decodedSequencePitchInBytes,
+    size_t encodedSequencePitchInInts,
     int numSequences
 ){
     auto group = cg::tiled_partition<groupsize>(cg::this_thread_block());
@@ -699,8 +699,8 @@ void callDecodeSequencesFrom2BitKernel(
     char* d_decodedSequences,
     const unsigned int* d_encodedSequences,
     const int* d_sequenceLengths,
-    int decodedSequencePitchInBytes,
-    int encodedSequencePitchInInts,
+    size_t decodedSequencePitchInBytes,
+    size_t encodedSequencePitchInInts,
     int numSequences,
     int groupsize,
     cudaStream_t stream
